Support for preserving files in the destination dir for Sync (#660)
This is a gap from the Ant task, which supports preserving files
via preserveInTarget. This brings that functionality to the Sync
task to avoid having to rely on Ant.

+++ subprojects/core/src/main/java/org/gradle/api/internal/file/copy/SyncCopyActionDecorator.java
@@ -15,6 +15,12 @@
  */
 package org.gradle.api.internal.file.copy;
 
+import java.io.File;
+import java.util.HashSet;
+import java.util.Set;
+
+import org.gradle.api.file.CopySpec;
+import org.gradle.api.file.FileTreeElement;
 import org.gradle.api.file.FileVisitDetails;
 import org.gradle.api.file.FileVisitor;
 import org.gradle.api.file.RelativePath;
@@ -22,20 +28,24 @@ import org.gradle.api.internal.file.CopyActionProcessingStreamAction;
 import org.gradle.api.internal.file.collections.DirectoryFileTree;
 import org.gradle.api.internal.file.collections.MinimalFileTree;
 import org.gradle.api.internal.tasks.SimpleWorkResult;
+import org.gradle.api.specs.Spec;
 import org.gradle.api.tasks.WorkResult;
+import org.gradle.api.tasks.util.PatternSet;
 import org.gradle.util.GFileUtils;
-
-import java.io.File;
-import java.util.HashSet;
-import java.util.Set;
 
 public class SyncCopyActionDecorator implements CopyAction {
     private final File baseDestDir;
     private final CopyAction delegate;
+    private CopySpec preserveSpec;
 
     public SyncCopyActionDecorator(File baseDestDir, CopyAction delegate) {
+        this(baseDestDir, delegate, null);
+    }
+
+    public SyncCopyActionDecorator(File baseDestDir, CopyAction delegate, CopySpec preserveSpec) {
         this.baseDestDir = baseDestDir;
         this.delegate = delegate;
+        this.preserveSpec = preserveSpec;
     }
 
     public WorkResult execute(final CopyActionProcessingStream stream) {
@@ -52,7 +62,8 @@ public class SyncCopyActionDecorator implements CopyAction {
             }
         });
 
-        SyncCopyActionDecoratorFileVisitor fileVisitor = new SyncCopyActionDecoratorFileVisitor(visited);
+        SyncCopyActionDecoratorFileVisitor fileVisitor = new SyncCopyActionDecoratorFileVisitor(visited,
+            preserveSpec);
 
         MinimalFileTree walker = new DirectoryFileTree(baseDestDir).postfix();
         walker.visit(fileVisitor);
@@ -63,10 +74,21 @@ public class SyncCopyActionDecorator implements CopyAction {
 
     private static class SyncCopyActionDecoratorFileVisitor implements FileVisitor {
         private final Set<RelativePath> visited;
+        private final Spec<FileTreeElement> preserveSpec;
+        private final PatternSet preserveSet;
         private boolean didWork;
 
-        private SyncCopyActionDecoratorFileVisitor(Set<RelativePath> visited) {
+        private SyncCopyActionDecoratorFileVisitor(
+            Set<RelativePath> visited,
+            CopySpec preserveSpec) {
             this.visited = visited;
+            PatternSet preserveSet = new PatternSet();
+            if (preserveSpec != null) {
+                preserveSet.include(preserveSpec.getIncludes());
+                preserveSet.exclude(preserveSpec.getExcludes());
+            }
+            this.preserveSet = preserveSet;
+            this.preserveSpec = preserveSet.getAsSpec();
         }
 
         public void visitDir(FileVisitDetails dirDetails) {
@@ -80,12 +102,14 @@ public class SyncCopyActionDecorator implements CopyAction {
         private void maybeDelete(FileVisitDetails fileDetails, boolean isDir) {
             RelativePath path = fileDetails.getRelativePath();
             if (!visited.contains(path)) {
-                if (isDir) {
-                    GFileUtils.deleteDirectory(fileDetails.getFile());
-                } else {
-                    GFileUtils.deleteQuietly(fileDetails.getFile());
+                if (preserveSet.isEmpty() || !preserveSpec.isSatisfiedBy(fileDetails)) {
+                    if (isDir) {
+                        GFileUtils.deleteDirectory(fileDetails.getFile());
+                    } else {
+                        GFileUtils.deleteQuietly(fileDetails.getFile());
+                    }
+                    didWork = true;
                 }
-                didWork = true;
             }
         }
     }
